requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.3
djangorestframework==3.12.4
idna==2.10
pycparser==2.20
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
zope.event==4.5.0
zope.interface==5.4.0
